requirements.txt
# Chatterbox Streaming Package for TRUE model-level streaming
# This package includes BOTH generate() and generate_stream() methods
# NOTE: Multilingual support is not available - English only
chatterbox-streaming @ git+https://github.com/davidbrowne17/chatterbox-streaming.git

# OLD: Removed due to numpy version conflict with chatterbox-streaming
# chatterbox-tts @ git+https://github.com/travisvn/chatterbox-multilingual.git@exp

# Required for Chatterbox TTS watermarker functionality
resemble-perth

# PyTorch with CPU support (ensure CPU compatibility)
torch>=2.0.0,<2.7.0
torchaudio>=2.0.0,<2.7.0

# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# WebSocket support for real-time streaming
websockets>=12.0

# Required for FastAPI file upload support
python-multipart>=0.0.6

# Server-Sent Events support for real-time progress updates
sse-starlette>=3.0.2

# Environment variables
python-dotenv>=1.0.0

# Memory monitoring
psutil>=5.9.0

# Audio processing for long text concatenation
pydub>=0.25.1

# Testing Dependencies
requests>=2.28.0

# Required for perth watermarker
setuptools>=65.0.0 